Fixed dependency issue and some string issues.
I fixed the build path to bundle the android support library into the
apk package, instead of statically linking it as before. Also I adjusted
the string the_caf to escape the quotes, and also the string
title_activity_generating_string so that the ellipses is displayed as a
single character.